+++ packages/discovery/src/discovery/source/SourceCodeService.ts
@@ -5,6 +5,7 @@ import { DiscoveryProvider } from '../provider/DiscoveryProvider'
 import { deduplicateAbi } from './deduplicateAbi'
 import { getLegacyDerivedName } from './getDerivedName'
 import { processSources } from './processSources'
+import { skipIgnoredFunctions } from './skipIgnoredFunctions'
 
 export interface ContractSources {
   name: string
@@ -44,5 +45,25 @@ export class SourceCodeService {
     const isVerified = metadata.every((x) => x.isVerified)
 
     return { name, isVerified, abi, abis, files }
+  }
+
+  getRelevantAbi(
+    abis: Record<string, string[]>,
+    address: EthereumAddress,
+    implementations?: EthereumAddress[],
+    ignoreInWatchMode?: string[],
+  ): string[] {
+    const addresses = [address, ...(implementations ?? [])]
+    const relevantAbis = addresses.flatMap((add) => {
+      const abiEntry = Object.entries(abis).find(
+        ([key]) => key === add.toString(),
+      )
+      return abiEntry ? abiEntry[1] : []
+    })
+
+    const abi = deduplicateAbi(relevantAbis)
+    const relevantAbi = skipIgnoredFunctions(abi, ignoreInWatchMode)
+
+    return relevantAbi
   }
 }