+++ packages/discovery/src/discovery/source/skipIgnoredFunctions.ts
@@ -1,0 +1,22 @@
+import { makeFragment } from './deduplicateAbi'
+
+export function skipIgnoredFunctions(
+  abi: string[],
+  ignoreInWatchMode?: string[],
+): string[] {
+  if (!ignoreInWatchMode) {
+    return abi
+  }
+  const relevantAbi = []
+  for (const entry of abi) {
+    const fragment = makeFragment(entry)
+    if (!fragment) {
+      continue
+    }
+    const name = fragment.name
+    if (!ignoreInWatchMode.includes(name)) {
+      relevantAbi.push(entry)
+    }
+  }
+  return relevantAbi
+}